fix load aws secrets

+++ kubernetes/install.sh
@@ -8,13 +8,25 @@ while [ $# -gt 0 ]; do
   --nsp=*)
     nsp="${1#*=}"
     ;;
+  --aws_region_name=*)
+    aws_region_name="${1#*=}"
+    ;;
+  --aws_access_key=*)
+    aws_access_key="${1#*=}"
+    ;;
+  --aws_secret_key=*)
+    aws_secret_key="${1#*=}"
+    ;;
+  --aws_secret_name=*)
+    aws_secret_name="${1#*=}"
+    ;;
   --dry_run=*)
     dry_run="${1#*=}"
     ;;
   *)
-    printf "***************************\n"
-    printf "* Invalid argument (${1}).\n"
-    printf "***************************\n"
+    echo "***************************\n"
+    echo "* Invalid argument (${1}).\n"
+    echo "***************************\n"
     exit 1
     ;;
   esac
@@ -23,16 +35,31 @@ done
 
 name=${name:-"bitlipa-api"}
 nsp=${nsp:-"default"}
-region_name=${region_name:-"eu-central-1"}
-secret_name=${secret_name:-"arn:aws:secretsmanager:eu-central-1:931829732782:secret:bitlipa-api-secrets-Cfnp2a"}
+aws_region_name=${aws_region_name:-"eu-central-1"}
+aws_secret_name=${aws_secret_name:-"arn:aws:secretsmanager:eu-central-1:931829732782:secret:bitlipa-api-secrets-Cfnp2a"}
+
+if [[ "$aws_access_key" == "" ]] || [[ "$aws_secret_key" == "" ]]; then
+  echo "***************************\n"
+  echo "--aws_access_key and --aws_secret_key are required"
+  echo "***************************\n"
+  exit 1
+fi
 
 cd $(dirname "$0")/bitlipa-api/
 
 if [[ "$dry_run" == "y" ]] || [[ "$dry_run" == "yes" ]]; then
-  helm install --dry-run --set regionName=$region_name --set secretName=$secret_name --debug $name . -n $nsp
+  helm install --dry-run \
+    --set AWS_REGION_NAME=$aws_region_name \
+    --set AWS_ACCESS_KEY=$aws_access_key \
+    --set AWS_SECRET_KEY=$aws_secret_key \
+    --set AWS_SECRET_NAME=$aws_secret_name --debug $name . -n $nsp
 else
   helm delete $name -n $nsp
-  helm install --set regionName=$region_name --set secretName=$secret_name $name . -n $nsp
+  helm install \
+    --set AWS_REGION_NAME=$aws_region_name \
+    --set AWS_ACCESS_KEY=$aws_access_key \
+    --set AWS_SECRET_KEY=$aws_secret_key \
+    --set AWS_SECRET_NAME=$aws_secret_name $name . -n $nsp
 fi
 
 SERVICE_IP=$(kubectl get svc --namespace $nsp $name --template "{{ range (index .status.loadBalancer.ingress 0) }}{{.}}{{ end }}")
